Add missing `@octokit/request-error` to `package.json`
`src/main.ts` includes the import:
```javascript
import { RequestError } from '@octokit/request-error'
```

However, we weren't explicitly requiring this in `package.json`.
It was implicitly coming in via `@actions/github` import, but best
to be explicit about it.

Discovered via:
```shell
npm install -g npm-check
npm-check
```

Which reported the following error:
```
@octokit/request-error             😟  PKG ERR!  Not in the package.json. Found in: /src/main.test.ts, /src/main.ts
```

Note: There is a _much_ newer version of `@octokit/request-error` available.
However, that threw some type errors due to breaking changes.
This `2.1.0` version was already pinned as a transitive dependency
in `package-lock.json`, so I went with that for now.