Make typescript compile to `"es2022"` (#609)
Keeping back in `"es6"` is causing an issue over in a Dependabot PR where
another library wants to use newer Javascript features.

The reason `"es6"` was the previous target is because GitHub actions
used to use node 12, which didn't support some of the newer versions.
But now that GitHub Actions is on Node 20, which supports `"es2022"`,
it's safe to let Typescript compile using this much newer syntax.